requirements.txt
sentence-transformers==3.4.1
llama-cpp-python>=0.3.2
chromadb==0.6.3
fastapi==0.115.6
uvicorn[standard]==0.34.0
typer==0.15.1
markdown-it-py==3.0.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
